+++ src/test/java/org/ssutt/core/testing/TestDF.java
@@ -22,7 +22,7 @@ import org.junit.Assert;
 import org.junit.FixMethodOrder;
 import org.junit.Test;
 import org.junit.runners.MethodSorters;
-import org.ssutt.core.fetch.TTDataFetcher;
+import org.ssutt.core.fetch.AbstractDataFetcher;
 import org.ssutt.core.fetch.SSUDataFetcher;
 
 import java.util.ArrayList;
@@ -33,7 +33,7 @@ import java.util.TreeSet;
 @FixMethodOrder(MethodSorters.NAME_ASCENDING)
 public class TestDF {
 
-    private TTDataFetcher df;
+    private AbstractDataFetcher df;
 
     @Test
     public void aTestGettingDataFromSSU() {